+++ packages/localizations/src/ar-SA.ts
@@ -460,6 +460,7 @@ export const arSA: LocalizationResource = {
       actionText__join_waitlist: undefined,
       subtitle: 'للمتابعة إلى {{applicationName}}',
       title: 'تسجيل الدخول',
+      __experimental_titleCombined: undefined,
     },
     totpMfa: {
       formTitle: 'رمز التحقق',