+++ packages/localizations/src/pl-PL.ts
@@ -460,6 +460,7 @@ export const plPL: LocalizationResource = {
       actionText__join_waitlist: undefined,
       subtitle: 'aby przejść do {{applicationName}}',
       title: 'Zaloguj się',
+      __experimental_titleCombined: undefined,
     },
     totpMfa: {
       formTitle: 'Kod weryfikacyjny',