+++ packages/nextjs/src/server/utils.ts
@@ -3,7 +3,7 @@ import { constants } from '@clerk/backend/internal';
 import { isDevelopmentFromSecretKey } from '@clerk/shared/keys';
 import { logger } from '@clerk/shared/logger';
 import { isHttpOrHttps } from '@clerk/shared/proxy';
-import { handleValueOrFn } from '@clerk/shared/utils';
+import { handleValueOrFn, isProductionEnvironment } from '@clerk/shared/utils';
 import AES from 'crypto-js/aes';
 import encUtf8 from 'crypto-js/enc-utf8';
 import hmacSHA1 from 'crypto-js/hmac-sha1';
@@ -97,7 +97,7 @@ export function decorateRequest(
   req: ClerkRequest,
   res: Response,
   requestState: RequestState,
-  requestData?: AuthenticateRequestOptions,
+  requestData: AuthenticateRequestOptions,
 ): Response {
   const { reason, message, status, token } = requestState;
   // pass-through case, convert to next()
@@ -195,7 +195,7 @@ export function assertAuthStatus(req: RequestLike, error: string) {
   }
 }
 
-export function assertKey(key: string, onError: () => never): string {
+export function assertKey(key: string | undefined, onError: () => never): string {
   if (!key) {
     onError();
   }
@@ -224,6 +224,8 @@ export function assertTokenSignature(token: string, key: string, signature?: str
   }
 }
 
+const KEYLESS_ENCRYPTION_KEY = 'clerk_keyless_dummy_key';
+
 /**
  * Encrypt request data propagated between server requests.
  * @internal
@@ -233,7 +235,7 @@ export function encryptClerkRequestData(requestData?: Partial<AuthenticateReques
     return;
   }
 
-  if (requestData.secretKey && !ENCRYPTION_KEY) {
+  if (requestData.secretKey && !ENCRYPTION_KEY && isProductionEnvironment()) {
     // TODO SDK-1833: change this to an error in the next major version of `@clerk/nextjs`
     logger.warnOnce(
       'Clerk: Missing `CLERK_ENCRYPTION_KEY`. Required for propagating `secretKey` middleware option. See docs: https://clerk.com/docs/references/nextjs/clerk-middleware#dynamic-keys',
@@ -242,10 +244,11 @@ export function encryptClerkRequestData(requestData?: Partial<AuthenticateReques
     return;
   }
 
-  return AES.encrypt(
-    JSON.stringify(requestData),
-    ENCRYPTION_KEY || assertKey(SECRET_KEY, () => errorThrower.throwMissingSecretKeyError()),
-  ).toString();
+  const maybeKeylessEncryptionKey = isProductionEnvironment()
+    ? ENCRYPTION_KEY || assertKey(SECRET_KEY, () => errorThrower.throwMissingSecretKeyError())
+    : ENCRYPTION_KEY || SECRET_KEY || KEYLESS_ENCRYPTION_KEY;
+
+  return AES.encrypt(JSON.stringify(requestData), maybeKeylessEncryptionKey).toString();
 }
 
 /**
@@ -259,8 +262,12 @@ export function decryptClerkRequestData(
     return {};
   }
 
+  const maybeKeylessEncryptionKey = isProductionEnvironment()
+    ? ENCRYPTION_KEY || SECRET_KEY
+    : ENCRYPTION_KEY || SECRET_KEY || KEYLESS_ENCRYPTION_KEY;
+
   try {
-    const decryptedBytes = AES.decrypt(encryptedRequestData, ENCRYPTION_KEY || SECRET_KEY);
+    const decryptedBytes = AES.decrypt(encryptedRequestData, maybeKeylessEncryptionKey);
     const encoded = decryptedBytes.toString(encUtf8);
     return JSON.parse(encoded);
   } catch (err) {