+++ packages/ui/src/descriptors.ts
@@ -1,4 +1,6 @@
-import { OAUTH_PROVIDERS, type OAuthProvider, WEB3_PROVIDERS, type Web3Provider } from '@clerk/types';
+import { OAUTH_PROVIDERS } from '@clerk/shared/oauth';
+import { WEB3_PROVIDERS } from '@clerk/shared/web3';
+import type { OAuthProvider, Web3Provider } from '@clerk/types';
 
 export const DESCRIPTORS = [
   // Alert